PlayList App using Binary tree adt C++ 2014
Written in C++ in 2014. This program loads a list of songs from an
external file and stores them in a binary tree adt. Each song includes a
title, an artist, an album name, the number
of minutes, and how many
times it has been played. The user can enter a title and retrieve all
songs in the tree with that title. The user can also delete a song by
title.